Use a companion object rather than an invalid package name
